requirements.txt
flask>=2.2.0
requests>=2.28.0
python-dotenv>=1.0.0
slack_sdk>=3.21.0
